{{cookiecutter.project_slug}}/src/requirements/base.txt
fastapi==0.100.0
pydantic==2.0.2
gunicorn==20.1.0
uvicorn[standard]==0.22.0
alembic==1.11.1
asyncpg==0.28.0
SQLAlchemy==2.0.18
pydantic-settings==2.0.2
passlib[bcrypt]==1.7.0
pyjwt==2.4.0